# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Planton — Plataforma de Plantões Médicos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/recommend-shifts:
    post:
      operationId: recommendShifts
      tags: [ai]
      summary: Recomendar o plantão ideal a partir de especialidade e CEP
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationInput"
      responses:
        "200":
          description: Até três recomendações com dados reais dos plantões
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
        "400":
          description: Nome, especialidade ou CEP inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Serviço de recomendação indisponível
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Shifts ─────────────────────────────────────────────────────────────────
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: Candidaturas do usuário logado
      responses:
        "200":
          description: Lista de candidaturas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"

    post:
      operationId: createApplication
      tags: [applications]
      summary: Candidatar-se a um plantão (médico)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Candidatura criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Dados inválidos ou já candidatado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{id}:
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Atualizar status de candidatura (hospital seleciona/rejeita)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Candidatura atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Candidatura não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── User ───────────────────────────────────────────────────────────────
...
    Shift:
      type: object
      required: [id, hospitalId, hospitalName, title, specialty, date, startTime, endTime, remuneration, address, city, state, status, createdAt]
      properties:
        id:
          type: number
        hospitalId:
          type: number
        hospitalName:
          type: string
        title:
          type: string
        specialty:
          type: string
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        remuneration:
          type: number
          minimum: 0.01
        address:
          type: string
        city:
          type: string
        state:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        status:
          type: string
          enum: [ABERTO, PREENCHIDO, CANCELADO, ENCERRADO]
        applicationsCount:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
...
    RecommendationInput:
      type: object
      required: [name, specialty, cep]
      properties:
        name:
          type: string
          minLength: 2
        specialty:
          type: string
          minLength: 2
        cep:
          type: string
          pattern: '^\d{8}$'
    ShiftRecommendation:
      type: object
      required: [shiftId, score, reason, shift]
      properties:
        shiftId:
          type: number
        score:
          type: number
          minimum: 0
          maximum: 100
        reason:
          type: string
        distanceKm:
          type: ["number", "null"]
        shift:
          $ref: "#/components/schemas/Shift"
    RecommendationResponse:
      type: object
      required: [recommendations, cached]
      properties:
        recommendations:
          type: array
          maxItems: 1
          items:
            $ref: "#/components/schemas/ShiftRecommendation"
        cached:
          type: boolean
        message:
          type: string

    # ── Application ────────────────────────────────────────────────────────
    Application:
      type: object
      required: [id, shiftId, doctorId, status, createdAt]
      properties:
        id:
          type: number
        shiftId:
          type: number
        doctorId:
          type: number
        doctorName:
          type: ["string", "null"]
        doctorSpecialty:
          type: ["string", "null"]
        doctorCrm:
          type: ["string", "null"]
        shift:
          $ref: "#/components/schemas/Shift"
        status:
          type: string
          enum: [PENDENTE, SELECIONADO, REJEITADO]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        payment:
          oneOf:
            - $ref: "#/components/schemas/PaymentSummary"
            - type: "null"
    PaymentSummary:
      type: object
      required: [id, status, grossAmount, platformFeePercent, platformFeeAmount, doctorAmount]
      properties:
        id:
          type: number
        status:
          type: string
          enum: [AGUARDANDO_PAGAMENTO, PAGO, LIBERADO, ESTORNADO, FALHOU]
        asaasStatus:
          type: ["string", "null"]
        invoiceUrl:
          type: ["string", "null"]
        grossAmount:
          type: number
        platformFeePercent:
          type: number
        platformFeeAmount:
          type: number
        doctorAmount:
          type: number
        paidAt:
          type: ["string", "null"]
        releasedAt:
          type: ["string", "null"]
    ApplicationInput:
      type: object
      required: [shiftId]
      properties:
        shiftId:
          type: number
        notes:
          type: string
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [PENDENTE, SELECIONADO, REJEITADO]
        notes:
          type: string