# lib/api-zod/src/generated/types/paymentSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Planton — Plataforma de Plantões Médicos
 * OpenAPI spec version: 1.0.0
 */
import type { PaymentSummaryStatus } from './paymentSummaryStatus';

export interface PaymentSummary {
  id: number;
  status: PaymentSummaryStatus;
  /** @nullable */
  asaasStatus?: string | null;
  /** @nullable */
  invoiceUrl?: string | null;
  grossAmount: number;
  platformFeePercent: number;
  platformFeeAmount: number;
  doctorAmount: number;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  releasedAt?: string | null;
}